adapt project structure to some recommendations
Of course, separating test from source is often a good thing. Nobody though
tells you how cumbersome it is to get your code loaded by the tests with pytest..

Web is full of people wondering and complaining. In the end, of course, python;s
PYTHONPATH must be set to our project src folder so that the code is
found as a module. No magic.